break toolbar out into new component
